requirements.txt
# Python==3.7.10
Flask==1.1.2
Werkzeug==1.0.1
flask-blueprint==1.3.0
Flask-Cors==3.0.9
gunicorn==20.0.4
gevent==21.1.2
numpy
opencv-python
matplotlib
pillow
pandas
psycopg2
dataframe_image
openpyxl
virtualenv
flask_migrate
Flask-SQLAlchemy